+++ module/combat-selector/combatselector.js
@@ -12,13 +12,13 @@ Hooks.on("targetToken", (user, token, targeted) => {
 // When an application is rendered, check if the automated defense is applicable
 Hooks.on("renderApplication", (app, html, data) => {
   // Get the text and check it for the weapon dicepools
-  let textCheck = html[0].innerText
+  const textCheck = html[0].innerText
   if(textCheck.includes("Strength + Brawl") || textCheck.includes("Strength + Weaponry") || textCheck.includes("Dexterity + Athletics")){
     // Check the game settings if the penalty should be applied automatically
     if(game.settings.get("cofdutils", "combatselector-automatedDefense")){
       // Grab the currently selected token's defense, reverse it, and then add it to the existing bonus/penalty.
-      let targetedDefense = -Math.abs(game.settings.get("cofdutils", "combatselector-penalty"))
-      let totalPenalty = targetedDefense + data.bonusDice
+      const targetedDefense = -Math.abs(game.settings.get("cofdutils", "combatselector-penalty"))
+      const totalPenalty = Number(targetedDefense) + Number(data.bonusDice)
       
       // Change the input field on the weapon dialogue
       html.find("input[name='dicePoolBonus']").attr("value", totalPenalty)