+++ module/group-beats/groupbeats.js
@@ -11,6 +11,11 @@ export class GroupBeats extends FormApplication {
       resizable: true,
       minimizable: true,
       closeOnSubmit: false,
+      tabs: [{
+        navSelector: ".sheet-tabs",
+        contentSelector: ".sheet-body",
+        initial: "players"
+      }],
       dragDrop: [{
         dragSelector: ".entity.actor",
         dropSelector: null
@@ -27,19 +32,34 @@ export class GroupBeats extends FormApplication {
   
   // Send data to the template
   getData() {
-    // Remove actors that have been deleted
-    this.activePlayers.forEach(actor => {      
-      this.isActorAlive(actor.id)
+    const data = super.getData()
+    
+    // Define some Handlebars variables
+    data.currentBeats = this.currentBeats
+    data.playerList = []
+    
+    // Re-render the actor's data each time the data is called for
+    this.activePlayers.forEach(actorID => {
+      if(this.isActorAlive(actorID)){
+        // Get the actor's sheet
+        const player = game.actors.get(actorID)
+                
+        // Push only the data we need and format it nicely for Handlebar rendering
+        data.playerList.push({
+          id: player.data.id,
+          img: player.data.img,
+          name: player.data.name,
+          beats: (player.data.data.beats % 5),
+          xp: ~~(player.data.data.beats/5)
+        })
+      }
     })
     
-    return {
-      currentBeats: this.currentBeats,
-      activePlayers: this.activePlayers
-    }
+    return data
   }
 
   // When the form is updated, re-render the template
-  async _updateObject(event, formData) {
+  _updateObject(event, formData) {
     this.render()
   }
   
@@ -52,21 +72,23 @@ export class GroupBeats extends FormApplication {
     html.find(".distributeBeats").click(this._distributeBeats.bind(this))
     html.find(".dropBeats").click(this._dropTheBeats.bind(this))
     html.find(".dropActors").click(this._dropActors.bind(this))
+    html.find(".playerImage").click(this._setSelected.bind(this))
   }
   
   // Response to an actor being dropped onto the beats menu
-  _onDrop(event) {
+  _onDrop(event) {    
     let data
+    
     try {
-        data = JSON.parse(event.dataTransfer.getData('text/plain'))
+      data = JSON.parse(event.dataTransfer.getData('text/plain'))
     }
     catch (err) {
-        return false
+      return false
     }
 
 
     if (data.type == 'Actor') {
-      this.addActor(data)
+      this.addActor(data.id)
     }
   }
   
@@ -80,19 +102,19 @@ export class GroupBeats extends FormApplication {
         // Update the game settings and the beatsMenu data, then re-render
         game.settings.set("cofdutils", "groupbeats-currentBeats", newBeats)
         this.currentBeats = newBeats
-        this.render(true)
       } else {
         // Otherwise, just add the new beats in
         
         // Define variables
-        let currentBeats = this.currentBeats
-        let newBeatsTotal = currentBeats + newBeats
+        const currentBeats = this.currentBeats
+        const newBeatsTotal = currentBeats + newBeats
         
         // Update the game settings and the beatsMenu data, then re-render
         game.settings.set("cofdutils", "groupbeats-currentBeats", newBeatsTotal)
         this.currentBeats = newBeatsTotal
-        this.render(true)
       }
+      
+      this.render()
     }
   }
   
@@ -127,7 +149,7 @@ export class GroupBeats extends FormApplication {
     // Update the game settings with the updated list
     game.settings.set("cofdutils", "groupbeats-activePlayers", this.activePlayers)
     // Re-render the page
-    this.render(true)
+    this.render()
   }
   
   // Adding new beats function
@@ -150,13 +172,13 @@ export class GroupBeats extends FormApplication {
            label: `${game.i18n.localize("CofD.GroupBeatsMenu.addBeats")}`,
            callback: (html) => {           
              // Define the new number of beats, and force it to be a number
-             let newBeats = Number(html.find('#newBeats')[0].value)
+             const newBeats = Number(html.find('#newBeats')[0].value)
              
              // Set the new number of beats in the module
              this.updateBeats(newBeats)
              
              // Define the chat message to output
-             let message = `${game.i18n.localize("CofD.GroupBeatsMenu.Message.added")} ${newBeats} ${game.i18n.localize("CofD.GroupBeatsMenu.Message.newBeats")} ${this.currentBeats}`
+             const message = `${game.i18n.localize("CofD.GroupBeatsMenu.Message.added")} ${newBeats} ${game.i18n.localize("CofD.GroupBeatsMenu.Message.newBeats")} ${this.currentBeats}`
                
              // Generate a chat message to notify everyone that story beats have been added
              ChatMessage.create({
@@ -182,60 +204,80 @@ export class GroupBeats extends FormApplication {
     // Distribute beats function
     _distributeBeats() {
       // set of players to distribute.
-      let actors = this.activePlayers
+      const actors = this.activePlayers
 
       if(actors.length > 0){
         // Define how many beats each player gets.
-        let beatsPerPlayer = ~~(this.currentBeats / actors.length) // Whole number
-        let remainderBeats = (this.currentBeats % actors.length) // Remainder
+        const beatsPerPlayer = ~~(this.currentBeats / actors.length) // Whole number
+        const remainderBeats = (this.currentBeats % actors.length) // Remainder
         
         // Start of a chat message for awarding beats.
         let msg = `
           <b>${beatsPerPlayer} ${game.i18n.localize("CofD.GroupBeatsMenu.Message.beatsAwarded")}</b> 
           <br>${game.i18n.localize("CofD.GroupBeatsMenu.Message.recipients")}: 
         `
-            
-        actors.forEach(actor => 
-        {
-          // Check if the actors exist still, because if they don't this could
-          // cause an error
-          if(this.isActorAlive(actor.id)){
-            // Define variables for updating the beats
-            // Player/Actor stuff
-            let player = this.getEntity({ type: "Actor", id: actor.id })
-            let playerSheet = this.getActorSheet(actor.id)
-            // Beat numbers
-            let playerOldBeats = player.entity.data.data.beats
-            let playerNewBeats = Number(playerOldBeats) + Number(beatsPerPlayer)
-            
-            // Update the number of beats in the sheet
-            playerSheet.update({"data.beats": playerNewBeats})
-            
-            // Add each actor's name to the chat message
-            msg += `${player.entity.data.name}, `
-          }
+        
+        // Promise chain to keep everything in sane order
+        const promiseMeBeats = new Promise((resolve, reject) => {
+          actors.forEach(actorID => 
+          {
+            // Check if the actors exist still, because if they don't this could
+            // cause an error
+            if(this.isActorAlive(actorID)){
+              // Define variables for updating the beats
+              // Player/Actor stuff
+              const playerSheet = game.actors.get(actorID)
+              
+              // Beat numbers
+              const playerOldBeats = playerSheet.data.data.beats
+              const playerNewBeats = Number(playerOldBeats) + Number(beatsPerPlayer)
+              
+              // Update the beats in the actor's sheet
+              game.actors.get(actorID).update({"data.beats": playerNewBeats})
+              
+              // Add each actor's name to the chat message
+              msg += `${game.actors.get(actorID).data.name}, `
+            }
+          })
+          
+          // Wrap up crafting the chat message by removing the last ", " and replacing it with a period
+          msg = msg.substring(0, msg.length-2)
+          msg += '.'
+          
+          resolve()
         })
         
-        // Set the new number of beats to the remainder
-        this.updateBeats(remainderBeats, true)
-        
-        // Wrap up crafting the chat message
-        msg = msg.substring(0, msg.length-2)
-        msg += '.'
+        // Send an update to the beats
+        const promiseMeBeats2 = new Promise((resolve, reject) => {
+          this.updateBeats(remainderBeats, true)
+          
+          resolve()
+        })
         
         // Generate a chat message to notify everyone that story beats have been distributed
-        ChatMessage.create({
-          content: msg,
-          speaker: {
-            alias: game.i18n.localize("CofD.GroupBeatsMenu.alias")
-          }
-        }, {})
+        const promiseMeBeats3 = new Promise((resolve, reject) => {
+          ChatMessage.create({
+            content: msg,
+            speaker: {
+              alias: game.i18n.localize("CofD.GroupBeatsMenu.alias")
+            }
+          }, {})
+          
+          resolve()
+        })
+        
+        // Go through the promise chain
+        promiseMeBeats
+          .then(promiseMeBeats)
+          .then(promiseMeBeats2)
+          .then(promiseMeBeats3)
+          .then(console.log("CofDUtils BeatsMenu: Beats distributed to players!"))
       }
     }
     
     // Set the number of grroup beats to 0
     _dropTheBeats(){    
-      let message = game.i18n.localize("CofD.GroupBeatsMenu.Message.dropBeats")
+      const message = game.i18n.localize("CofD.GroupBeatsMenu.Message.dropBeats")
       
       // Set the number of beats to 0
       this.updateBeats(0, true)
@@ -254,74 +296,45 @@ export class GroupBeats extends FormApplication {
       this.updateActors("clear")
     }
     
-    // Get the actor sheet via ID
-    getActorSheet(id){
-      // Search through the game for the sheet
-      let sheetID = Object.keys(game.actors.contents).find(sheet => game.actors.contents[sheet].id === id)
-      
-      // Send the actor sheet back
-      return game.actors.contents[sheetID]
-    }
-    
-    // Get the actor entity
-    getEntity(data) {
-      // Retrieve the entity from the Actor collection
-      // Technically this works for all entities, but earlier functions
-      // prevent this from being called for anything but actors
-      let entity
-      entity = game.collections.get(data.type).get(data.id)
-      
-      // Easily storable entity format
-      let result = { entity: entity, data: {} }
-      
-      if (entity) {
-        // Store useful data
-        result.data = {
-          id: entity.id,
-          uuid: entity.uuid,
-          img: entity.img,
-          name: entity.name,
-          type: data.type
-        }
-      }
-      
-      // Send transformed entity data back
-      return result
+    // When a player icon is clicked, open up the player edit menu
+    _setSelected(player) {
+      // Get ID of the player clicked
+      const playerID = player.target.id
     }
     
     // Add a new actor to the active players list
-    addActor(data) {
+    addActor(actorID) {
       // Define the actor data
-      let actor = this.getEntity(data)
+      const actor = game.actors.get(actorID)
       
       // Check if the actor is unique in the already existing list;
       // Returns true if it's found, or false if it's not found
-      let actorUniqueCheck = this.activePlayers.find(players => players.id == actor.data.id)
-      
+      const actorUniqueCheck = this.activePlayers.find(players => players == actorID)
+            
       // If the actor exists and is unique
-      if (actor.entity && (!actorUniqueCheck)) {
+      if (!actorUniqueCheck) {
         // Push to the players list
-        this.updateActors("add", actor.data)
+        this.updateActors("add", actorID)
       }
     }
     
     // Checks if the actors still exist; if they were deleted, then remove them from the list
-    isActorAlive(id){
-      let actorAlive = Object.keys(game.actors.contents).find(sheet => game.actors.contents[sheet].id === id)
+    isActorAlive(actorID){
+      const actor = game.actors.get(actorID)
       
       // If the actor is dead, update the variables
-      if(actorAlive > -1){
+      if(actor){
         // Actor is alive, continue
         return true
+      } else {
+        // Actor is dead, rewrite some variables
+        const newPlayersList = this.activePlayers.filter((players) => players !== actorID)
+        
+        // Update players list
+        this.updateActors("replace", newPlayersList)
+        
+        // Continue
+        return false
       }
-      
-      // Actor is dead, rewrite some variables
-      let newPlayersList = this.activePlayers.filter((players) => players.id !== id)
-      
-      // Update players list
-      this.updateActors("replace", newPlayersList)
-      
-      // Continue
-      return false
     }
 }