+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+#### 0.1.5
+* Updated the Beats Menu display, now displaying Beats/XP and updating when the actors on the list update!
+* A bunch of code restructuring/fixes and cleaned up the way I get the actor sheets.
+* Changed a bunch of lets to consts because some variables were never changed.
+* Fix for how the Combat Selector displayed the dice mods.
+
 #### 0.1.4
 * Quick fix for applying automated defense to the weapon's dicepool.